mqtt-subscriber-mongodb/requirements.txt
dnspython==2.4.2
paho-mqtt==1.6.1
pymongo==4.4.1
pydantic==2.2.1
redis==5.0.0
python-dotenv==1.0.0